+++ strawtubes/strawtubes.cxx
@@ -25,6 +25,7 @@
 #include "FairRuntimeDb.h"
 #include "FairVolume.h"
 #include "ShipDetectorList.h"
+#include "ShipGeoUtil.h"
 #include "ShipStack.h"
 #include "TClonesArray.h"
 #include "TGeoBBox.h"
@@ -49,25 +50,6 @@ strawtubes::strawtubes(std::string medium)
 
 strawtubes::strawtubes(const char* name, Bool_t active)
     : Detector(name, active, kStraw) {}
-
-// -----   Private method InitMedium
-Int_t strawtubes::InitMedium(const char* name) {
-  static FairGeoLoader* geoLoad = FairGeoLoader::Instance();
-  static FairGeoInterface* geoFace = geoLoad->getGeoInterface();
-  static FairGeoMedia* media = geoFace->getMedia();
-  static FairGeoBuilder* geoBuild = geoLoad->getGeoBuilder();
-
-  FairGeoMedium* ShipMedium = media->getMedium(name);
-
-  if (!ShipMedium) {
-    Fatal("InitMedium", "Material %s not defined in media file.", name);
-    return -1111;
-  }
-  TGeoMedium* medium = gGeoManager->GetMedium(name);
-  if (medium != nullptr) return ShipMedium->getMediumIndex();
-
-  return geoBuild->createMedium(ShipMedium);
-}
 
 Bool_t strawtubes::ProcessHits(FairVolume* vol) {
   /** This method is called from the MC stepping */
@@ -206,15 +188,15 @@ void strawtubes::ConstructGeometry() {
       implement here you own way of constructing the geometry. */
 
   TGeoVolume* top = gGeoManager->GetTopVolume();
-  InitMedium("mylar");
+  ShipGeo::InitMedium("mylar");
   TGeoMedium* mylar = gGeoManager->GetMedium("mylar");
-  InitMedium("STTmix8020_1bar");
+  ShipGeo::InitMedium("STTmix8020_1bar");
   TGeoMedium* sttmix8020_1bar = gGeoManager->GetMedium("STTmix8020_1bar");
-  InitMedium("tungsten");
+  ShipGeo::InitMedium("tungsten");
   TGeoMedium* tungsten = gGeoManager->GetMedium("tungsten");
-  InitMedium(f_frame_material);
+  ShipGeo::InitMedium(f_frame_material);
   TGeoMedium* FrameMatPtr = gGeoManager->GetMedium(f_frame_material);
-  InitMedium(fMedium.c_str());
+  ShipGeo::InitMedium(fMedium.c_str());
   TGeoMedium* med = gGeoManager->GetMedium(fMedium.c_str());
 
   gGeoManager->SetVisLevel(4);
@@ -322,34 +304,30 @@ void strawtubes::ConstructGeometry() {
 
     for (Int_t vnb = 0; vnb < 4; vnb++) {
       // View loop
-      TString nmview;
-      Double_t angle;
-      Double_t stereo_growth;
-      Double_t stereo_pitch;
-      Double_t offset_layer;
-      Int_t straws_per_layer;
-
-      switch (vnb) {
-        case 0:
-          angle = 0.;
-          nmview = nmstation + "_y1";
-          break;
-        case 1:
-          angle = f_view_angle;
-          nmview = nmstation + "_u";
-          break;
-        case 2:
-          angle = -f_view_angle;
-          nmview = nmstation + "_v";
-          break;
-        case 3:
-          angle = 0.;
-          nmview = nmstation + "_y2";
-          break;
-        default:
-          angle = 0.;
-          nmview = nmstation + "_y1";
-      }
+      const Double_t angle = [&] {
+        switch (vnb) {
+          case 1:
+            return f_view_angle;
+          case 2:
+            return -f_view_angle;
+          default:
+            return 0.;
+        }
+      }();
+      const TString nmview = [&] {
+        switch (vnb) {
+          case 0:
+            return nmstation + "_y1";
+          case 1:
+            return nmstation + "_u";
+          case 2:
+            return nmstation + "_v";
+          case 3:
+            return nmstation + "_y2";
+          default:
+            return nmstation + "_y1";
+        }
+      }();
 
       // Adjustments in the stereo views
       // stereo_growth: extension of stereo views beyond aperture
@@ -357,13 +335,13 @@ void strawtubes::ConstructGeometry() {
       // offset_layer: layer offset in stereo views
       // straws_per_layer: number of straws in one layer with stereo extension
       // If angle == 0., all numbers return the case of non-stereo views.
-      stereo_growth =
+      const Double_t stereo_growth =
           TMath::Tan(TMath::Abs(angle) * TMath::Pi() / 180.0) * straw_length;
-      stereo_pitch =
+      const Double_t stereo_pitch =
           f_straw_pitch / TMath::Cos(TMath::Abs(angle) * TMath::Pi() / 180.0);
-      offset_layer =
+      const Double_t offset_layer =
           f_offset_layer / TMath::Cos(TMath::Abs(angle) * TMath::Pi() / 180.0);
-      straws_per_layer =
+      const Int_t straws_per_layer =
           std::ceil(2 * (f_aperture_height + stereo_growth) / stereo_pitch);
 
       for (Int_t lnb = 0; lnb < 2; lnb++) {
@@ -472,23 +450,18 @@ void strawtubes::StrawEndPoints(Int_t fDetectorID, TVector3& vbot,
   stat += statnb;
   stat += "_";
   stat += statnb;
-  TString view;
-  switch (vnb) {
-    case 0:
-      view = "_y1";
-      break;
-    case 1:
-      view = "_u";
-      break;
-    case 2:
-      view = "_v";
-      break;
-    case 3:
-      view = "_y2";
-      break;
-    default:
-      view = "_y1";
-  }
+  const TString view = [&] {
+    switch (vnb) {
+      case 1:
+        return TString("_u");
+      case 2:
+        return TString("_v");
+      case 3:
+        return TString("_y2");
+      default:
+        return TString("_y1");
+    }
+  }();
   TGeoNavigator* nav = gGeoManager->GetCurrentNavigator();
   TString prefix = "Tr";
   prefix += statnb;