+++ veto/veto.cxx
@@ -32,6 +32,7 @@
 #include "FairRuntimeDb.h"
 #include "FairVolume.h"
 #include "ShipDetectorList.h"
+#include "ShipGeoUtil.h"
 #include "ShipStack.h"
 #include "TClonesArray.h"
 #include "TGeoArb8.h"
@@ -743,25 +744,6 @@ TGeoVolume* veto::MakeSegments() {
   return tTankVol;
 }
 
-// -----   Private method InitMedium
-Int_t veto::InitMedium(const char* name) {
-  static FairGeoLoader* geoLoad = FairGeoLoader::Instance();
-  static FairGeoInterface* geoFace = geoLoad->getGeoInterface();
-  static FairGeoMedia* media = geoFace->getMedia();
-  static FairGeoBuilder* geoBuild = geoLoad->getGeoBuilder();
-
-  FairGeoMedium* ShipMedium = media->getMedium(name);
-
-  if (!ShipMedium) {
-    Fatal("InitMedium", "Material %s not defined in media file.", name);
-    return -1111;
-  }
-  TGeoMedium* medium = gGeoManager->GetMedium(name);
-  if (medium != nullptr) return ShipMedium->getMediumIndex();
-
-  return geoBuild->createMedium(ShipMedium);
-}
-
 // -------------------------------------------------------------------------
 /**
  * @brief Processes a hit in the veto detector.
@@ -842,11 +824,11 @@ void veto::PreTrack() {
 void veto::ConstructGeometry() {
   TGeoVolume* top = gGeoManager->GetTopVolume();
 
-  InitMedium("vacuums");
-  InitMedium("Aluminum");
-  InitMedium("helium");
-  InitMedium("Scintillator");
-  InitMedium("steel");
+  ShipGeo::InitMedium("vacuums");
+  ShipGeo::InitMedium("Aluminum");
+  ShipGeo::InitMedium("helium");
+  ShipGeo::InitMedium("Scintillator");
+  ShipGeo::InitMedium("steel");
 
   gGeoManager->SetNsegments(100);